New version of pageExtract

+++ pageExtract.php
@@ -1,11 +1,22 @@
 <?php
 include_once('init.php');
-getPages();
-function getPages() {
+
+getPages((isset($_GET['fullReset'])?1:0));
+function getPages($fullReset) {
 	set_time_limit(7200);
-	mysql_query("TRUNCATE wg_page");
-	mysql_query("TRUNCATE wg_links");
-	$r = mysql_query("SELECT * FROM wg_category WHERE distance>=1 ORDER BY id");
+	$startAt = 1;
+	if($fullReset == 1) {
+		mysql_query("TRUNCATE wg_page");
+		mysql_query("TRUNCATE wg_links");
+	}
+	else {
+		$r = mysql_query("SELECT * FROM wg_page ORDER BY category DESC LIMIT 1");
+		if($re = mysql_fetch_array($r)) {
+			$startAt = $re['category']; // restart where we left off
+		}
+	}
+	$r = mysql_query("SELECT * FROM wg_category WHERE distance>=1 AND killBranch=0 AND id>=$startAt ORDER BY id");
+	echo "SELECT * FROM wg_category WHERE distance>=1 AND killBranch=0 AND category>=$startAt ORDER BY id";
 	while($re = mysql_fetch_array($r)) {
 		extractPages($re['name'], $re);
 		echo $re['name']." is done<br />";
@@ -21,7 +32,7 @@ function extractPages($categoryName, $parent) {
 	if(!is_null($dom)) {
 		foreach ($dom->getElementsByTagName('a') as $link) {
 				$h = str_replace("/wiki/", "", $link->getAttribute('href')); $cleanName = wikiToName($h);
-				$p = mysql_query("SELECT * FROM wg_page WHERE name='$cleanName'");
+				$p = mysql_query("SELECT * FROM wg_page WHERE name=\"$cleanName\"");
 				if($pa = mysql_fetch_array($p)) {
 					// for now do nothing...
 				}
